Optional viewing, documentationf ully updated, various improvements

+++ src/foambryo/dcel.py
@@ -373,7 +373,7 @@ class DcelData:
 
     def __init__(self, points: NDArray[np.float64], triangles: NDArray[np.int64], labels: NDArray[np.int64]) -> None:
         """Take a multimaterial mesh as input."""
-        triangles_and_labels = np.hstack((triangles, labels))
+        triangles_and_labels = np.hstack((triangles, labels), dtype=np.int64)
         for i, f in enumerate(triangles_and_labels):
             if f[3] > f[4]:
                 triangles_and_labels[i] = triangles_and_labels[i, [0, 2, 1, 4, 3]]
